NEW: counterparties subscription. REFACTOR: custom resolvers separated to several files.
