+++ src/server/graphql/context.js
@@ -1,0 +1,10 @@
+// @flow
+
+import QBlockchainData from "../data/blockchain";
+import type { GraphQLRequestContext } from "../data/collection";
+
+export type GraphQLRequestContextEx = GraphQLRequestContext & {
+    data: QBlockchainData,
+}
+
+
